refactor(app): LocationConflictModal organisms
This is a pretty complex thing (could probably use some simplifying
refactors) so I think it deserves to be an organism.
