+++ app/src/organisms/LocationConflictModal/index.ts
@@ -1,0 +1,2 @@
+export * from './LocationConflictModal'
+export * from './ChooseModuleToConfigureModal'
